TEST(<tes_printf.c>): add test for printf.

+++ lib/ice/printf/private/ice_printf_ptr.c
@@ -14,7 +14,7 @@ bool ice_printf_ptr(buffer_t *buffer, va_list args)
 
     ASSERT_RET(!adds_buffer(buffer, "0x"), true);
 
-    ice_btoa(va_arg(args, unsigned int), str, "0123456789abcdef");
+    ice_btoa(va_arg(args, unsigned long int), str, "0123456789abcdef");
     ASSERT_RET(!add_width(buffer, str), true);
 
     return false;
